bastardkb/charybdis: use  instead of  for DPI functions

+++ keyboards/bastardkb/charybdis/charybdis.c
@@ -90,10 +90,10 @@ static void read_charybdis_config_from_eeprom(charybdis_config_t* config) {
 static void write_charybdis_config_to_eeprom(charybdis_config_t* config) { eeconfig_update_kb(config->raw); }
 
 /** \brief Return the current value of the pointer's default DPI. */
-static uint32_t get_pointer_default_dpi(charybdis_config_t* config) { return (uint32_t)config->pointer_default_dpi * CHARYBDIS_DEFAULT_DPI_CONFIG_STEP + CHARYBDIS_MINIMUM_DEFAULT_DPI; }
+static uint16_t get_pointer_default_dpi(charybdis_config_t* config) { return (uint16_t)config->pointer_default_dpi * CHARYBDIS_DEFAULT_DPI_CONFIG_STEP + CHARYBDIS_MINIMUM_DEFAULT_DPI; }
 
 /** \brief Return the current value of the pointer's sniper-mode DPI. */
-static uint32_t get_pointer_sniping_dpi(charybdis_config_t* config) { return (uint32_t)config->pointer_sniping_dpi * CHARYBDIS_SNIPING_DPI_CONFIG_STEP + CHARYBDIS_MINIMUM_SNIPING_DPI; }
+static uint16_t get_pointer_sniping_dpi(charybdis_config_t* config) { return (uint16_t)config->pointer_sniping_dpi * CHARYBDIS_SNIPING_DPI_CONFIG_STEP + CHARYBDIS_MINIMUM_SNIPING_DPI; }
 
 /** \brief Set the appropriate DPI for the input config. */
 static void maybe_update_pointing_device_cpi(charybdis_config_t* config) {
